Update WavefrontSpringBootAutoConfiguration.java
Part of the updates to support added URL to the .wavefront_token, as well as to also support ability for configuration to have one time link as its properties. This is needed in order for the outside application to know whether the application has just registered Wavefront, or is a revisiting user. Even though the wavefront.properties file may not exist (or setup), with the presence of .wavefront_token, springboot autoconfigure can function in such a way that you don't need to create a new freemium account everytime.