Starter v2
This commit restructure account management and auto-configuration so
that it reuses what Spring Boot and Micrometer do with a configured
account.

An EnvironmentPostProcessor checks if it needs to auto-negotiate an
account, or restore one from disk. A single-use login URL as well as
the information of the account is logged on startup.

An auto-configuration provides an `ApplicationTags` based on the
settings defined by `wavefront.application.*`. The instance can be
further customized, i.e. to add custom tags, by implementing one or more
ApplicationTagsBuilderCustomizer beans.

When metrics are exposed via Micrometer, the state of `ApplicationTags`
is used to automatically tag metrics.

When opentracing and the wavefront sdk is on the classpath, traces are
automatically exported to Wavefront. As Spring Boot now auto-configures
a `WavefrontSender` to export metrics, the auto-configured sender is
reused for traces as well. Traces can be disabled via
`wavefront.tracing.enabled=false`.

A sample is also available to demonstrate the starter. Starting the app
is all that's required to start exporting metrics to Wavefront.