Add actuator endpoint to Wavefront dashboard
This commit exposes an endpoint at `/actuator/wavefront` that redirects
to the Wavefront dashboard according to the current configuration.

If an account has been managed, a one time link is requested at every
invocation. If the account has been configured by the user, the
endpoint redirects to the Wavefront cluster.

Closes gh-16