Backfills tests for error cases and adds some notes
Those used to exclusively receiving data from OpenTracing apis are
probably less aware of other data practices that preceded or succeeded
it. This adds a bunch of notes (which I'll add to Brave for posterity).

Also, it backfills tests to
 * show not tell about various API practice
 * encourage using tests as a way of resolving misunderstanding

This also migrates the tests to the more straightforward Mvc controller
as there is bugs and scope glitches in some reactive instrumentation
which can result in loss of time with no gain when we are primarily
interested in data mapping here.