Rename ~/.wavefront_token to ~/.wavefront_freemium
This file stores the api token of a freemium account and is not supposed
to be used for a regular api token as we have no way to specify the URL
of the wavefront cluster.

To avoid users to consider they could use this file to share their api
token, this commit renames the file to make it clear it is about
freemium account only.

Closes gh-31