Disable account management in integration tests
Previously, running an integration test with the starter violated the
principle of least astonishment by negotiating an account and
configuring the context to export metrics to a real account.

This commit adds an extra check that verify if the current thread is
executing a test. If it does, account management is disabled and the
environment is left untouched. As side effect of that is that the test
might fail as an api-token is mandatory. However, this is more
consistent with other monitoring system and will make sure CI env do
not negotiate accounts that are not meant to be used.

Closes gh-56