Remove OpenTracing API from Wavefront starter
This commit excludes the OpenTracing dependency from the wavefront SDK
so that no tracing auto-configuration kicks in when the starter is
added.

Users can opt-in for either Spring Cloud Sleuth or OpenTracing which
will auto-configure the relevant bits to enable tracing.

This commit also makes sure that users that opt-in for Spring Cloud
Sleuth don't have OpenTracing libraries on their classpath.

JVM metrics are now exported independently, even when no tracing is
active. The relevant property has been moved to a `metrics` namespace
to make that more explicit.

Co-authored-by: Tommy Ludwig <tludwig@pivotal.io>