Apply consistent tags in Sleuth integration
Previously, when a service name isn't set, Sleuth exported span with a
"default" service tag, rather tha using the same default than metric
exports.

This commit removes the custom handling of  `@LocalServiceName` as the
auto-configuration for `ApplicationTags` already handled this anyway. If
the user configured a dedicated `ApplicationTags` bean, this takes
precedence as before.

This commit also tests the different scenario more explicitly.

This prevents "spring.zipkin.service.name" to actually set the name of
the service tags but this would be inconsistent with metrics export
anyway. The integration test has been updated to use
"spring.application.name" for clarity.

Closes gh-82