Always create WavefrontTracer in OpenTracing
Previously, when `WavefrontSender` bean does not exist, creation of
`WavefrontTracer` is skipped. This causes a problem where we want to
use `WavefrontTracer` but not send data to Wavefront via
`WavefrontSpanReporter`.

This commit changes to always create `WavefrontTracer` but
conditionally create `WavefrontSpanReporter` when `WavefrontSender`
bean exists; otherwise, composite reporter will be used.

To completely skip the creation of tracer, there is already
`wavefront.tracing.enabled` property available.